+++ boltdb_batch.go
@@ -19,38 +19,48 @@ func newBoltDBBatch(db *BoltDB) *boltDBBatch {
 	}
 }
 
-func (b *boltDBBatch) assertOpen() {
-	if b.ops == nil {
-		panic("batch has been written or closed")
-	}
-}
-
 // Set implements Batch.
-func (b *boltDBBatch) Set(key, value []byte) {
-	b.assertOpen()
+func (b *boltDBBatch) Set(key, value []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
+	if b.ops == nil {
+		return errBatchClosed
+	}
 	b.ops = append(b.ops, operation{opTypeSet, key, value})
+	return nil
 }
 
 // Delete implements Batch.
-func (b *boltDBBatch) Delete(key []byte) {
-	b.assertOpen()
+func (b *boltDBBatch) Delete(key []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if b.ops == nil {
+		return errBatchClosed
+	}
 	b.ops = append(b.ops, operation{opTypeDelete, key, nil})
+	return nil
 }
 
 // Write implements Batch.
 func (b *boltDBBatch) Write() error {
-	b.assertOpen()
+	if b.ops == nil {
+		return errBatchClosed
+	}
 	err := b.db.db.Batch(func(tx *bbolt.Tx) error {
 		bkt := tx.Bucket(bucket)
 		for _, op := range b.ops {
-			key := nonEmptyKey(nonNilBytes(op.key))
 			switch op.opType {
 			case opTypeSet:
-				if err := bkt.Put(key, op.value); err != nil {
+				if err := bkt.Put(op.key, op.value); err != nil {
 					return err
 				}
 			case opTypeDelete:
-				if err := bkt.Delete(key); err != nil {
+				if err := bkt.Delete(op.key); err != nil {
 					return err
 				}
 			}
@@ -61,8 +71,7 @@ func (b *boltDBBatch) Write() error {
 		return err
 	}
 	// Make sure batch cannot be used afterwards. Callers should still call Close(), for errors.
-	b.Close()
-	return nil
+	return b.Close()
 }
 
 // WriteSync implements Batch.
@@ -71,6 +80,7 @@ func (b *boltDBBatch) WriteSync() error {
 }
 
 // Close implements Batch.
-func (b *boltDBBatch) Close() {
+func (b *boltDBBatch) Close() error {
 	b.ops = nil
+	return nil
 }