+++ memdb_batch.go
@@ -1,6 +1,6 @@
 package db
 
-import "github.com/pkg/errors"
+import "fmt"
 
 // memDBBatch operations
 type opType int
@@ -32,27 +32,38 @@ func newMemDBBatch(db *MemDB) *memDBBatch {
 	}
 }
 
-func (b *memDBBatch) assertOpen() {
-	if b.ops == nil {
-		panic("batch has been written or closed")
-	}
-}
-
 // Set implements Batch.
-func (b *memDBBatch) Set(key, value []byte) {
-	b.assertOpen()
+func (b *memDBBatch) Set(key, value []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if value == nil {
+		return errValueNil
+	}
+	if b.ops == nil {
+		return errBatchClosed
+	}
 	b.ops = append(b.ops, operation{opTypeSet, key, value})
+	return nil
 }
 
 // Delete implements Batch.
-func (b *memDBBatch) Delete(key []byte) {
-	b.assertOpen()
+func (b *memDBBatch) Delete(key []byte) error {
+	if len(key) == 0 {
+		return errKeyEmpty
+	}
+	if b.ops == nil {
+		return errBatchClosed
+	}
 	b.ops = append(b.ops, operation{opTypeDelete, key, nil})
+	return nil
 }
 
 // Write implements Batch.
 func (b *memDBBatch) Write() error {
-	b.assertOpen()
+	if b.ops == nil {
+		return errBatchClosed
+	}
 	b.db.mtx.Lock()
 	defer b.db.mtx.Unlock()
 
@@ -63,13 +74,12 @@ func (b *memDBBatch) Write() error {
 		case opTypeDelete:
 			b.db.delete(op.key)
 		default:
-			return errors.Errorf("unknown operation type %v (%v)", op.opType, op)
+			return fmt.Errorf("unknown operation type %v (%v)", op.opType, op)
 		}
 	}
 
 	// Make sure batch cannot be used afterwards. Callers should still call Close(), for errors.
-	b.Close()
-	return nil
+	return b.Close()
 }
 
 // WriteSync implements Batch.
@@ -78,6 +88,7 @@ func (b *memDBBatch) WriteSync() error {
 }
 
 // Close implements Batch.
-func (b *memDBBatch) Close() {
+func (b *memDBBatch) Close() error {
 	b.ops = nil
+	return nil
 }